pyproject.toml
[tool.poetry]
name = "rl-mind-dataset"
version = "0.1.0"
description = ""
authors = ["Asr419 <asrthebest17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
plotly = "^5.18.0"
python-dotenv = "^1.0.0"
nbformat = "^5.9.2"
torch = "1.13.1"
pytest = "^7.4.3"
pyarrow = "^14.0.2"
gymnasium = "^0.29.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
